requirements.txt
torch==1.4.0
pytorch-ignite==0.4.0
numpy
wandb
tqdm
scikit-learn
